Calibrate joints.
