+++ packages/lekiwi_sim/lekiwi_sim/assets/so_arm100/so_arm100.xml
@@ -36,10 +36,10 @@
         <joint axis="0 1 0" range="-1.92 1.92"/>
       </default>
       <default class="Pitch">
-        <joint axis="1 0 0" range="-3.32 0.174"/>
+        <joint axis="1 0 0" range="-1.747 1.747"/>
       </default>
       <default class="Elbow">
-        <joint axis="1 0 0" range="-0.174 3.14"/>
+        <joint axis="1 0 0" range="-1.657 1.657"/>
       </default>
       <default class="Wrist_Pitch">
         <joint axis="1 0 0" range="-1.66 1.66"/>
@@ -48,7 +48,8 @@
         <joint axis="0 1 0" range="-2.79 2.79"/>
       </default>
       <default class="Jaw">
-        <joint axis="0 0 1" range="-0.174 1.75"/>
+        <joint axis="0 0 1" range="0. 0.6"/>
+
       </default>
       <default class="visual">
         <geom type="mesh" contype="0" conaffinity="0" density="0" group="2" material="white"/>
@@ -77,14 +78,14 @@
         <geom type="mesh" mesh="Rotation_Pitch" class="visual"/>
         <geom type="mesh" mesh="Rotation_Pitch_Motor" class="motor_visual"/>
         <geom type="mesh" mesh="Rotation_Pitch" class="collision"/>
-        <body name="Upper_Arm" pos="0 0.1025 0.0306" euler="1.57079 0 0">
+        <body name="Upper_Arm" pos="0 0.1025 0.0306" euler="0 0 0">
           <inertial pos="-1.72052e-05 0.0701802 0.00310545" quat="0.50104 0.498994 -0.493562 0.50632" mass="0.162409"
             diaginertia="0.000213312 0.000167164 7.01522e-05"/>
           <joint name="Pitch" class="Pitch"/>
           <geom type="mesh" mesh="Upper_Arm" class="visual"/>
           <geom type="mesh" mesh="Upper_Arm_Motor" class="motor_visual"/>
           <geom type="mesh" mesh="Upper_Arm" class="collision"/>
-          <body name="Lower_Arm" pos="0 0.11257 0.028" euler="-1.57079 0 0">
+          <body name="Lower_Arm" pos="0 0.11257 0.028" euler="0 0 0">
             <inertial pos="-0.00339604 0.00137796 0.0768007" quat="0.701995 0.0787996 0.0645626 0.704859"
               mass="0.147968" diaginertia="0.000138803 0.000107748 4.84242e-05"/>
             <joint name="Elbow" class="Elbow"/>
@@ -119,7 +120,7 @@
                         <camera name="wrist" mode="fixed" resolution="480 640" fovy="90" pos="0. 0.0 0.01" euler="3.14 0. -1.57" />
                     </body>
                 </body>
-                <body name="Moving_Jaw" pos="-0.0202 -0.0244 0" quat="1.34924e-11 -3.67321e-06 1 -3.67321e-06">
+                <body name="Moving_Jaw" pos="-0.0202 -0.0244 0" euler="0 3.14 -0.174">
                   <inertial pos="-0.00161745 -0.0303473 0.000449646" quat="0.696562 0.716737 -0.0239844 -0.0227026"
                     mass="0.0202444" diaginertia="1.11265e-05 8.99651e-06 2.99548e-06"/>
                   <joint name="Jaw" class="Jaw"/>
@@ -154,7 +155,7 @@
   </contact>
 
   <keyframe>
-    <key name="home" qpos="0 -1.57 1.57 1.57 -1.57 0" ctrl="0 -1.57 1.57 1.57 -1.57 0"/>
-    <key name="rest" qpos="0 -3.32 3.11 1.18 0 -0.174" ctrl="0 -3.32 3.11 1.18 0 -0.174"/>
+    <key name="home" qpos="0 0 0 1.57 -1.57 0" ctrl="0 0 0 1.57 -1.57 0"/>
+    <key name="rest" qpos="0 -1.75 1.59 1.13 0 0" ctrl="0 -1.75 1.59 1.13 0 0"/>
   </keyframe>
 </mujoco>
